// src/schemas/games_db.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * Full name of the game which might includes tagged data
 */
export type Unknown = string;
/**
 * Short name of the game which might not be unique
 */
export type Unknown1 = string;
/**
 * Region of the game
 */
export type Unknown2 = string;
/**
 * Languages of the game, either as an array or a single string.
 */
export type Unknown3 = Unknown4 | Unknown5;
export type Unknown4 = string;
export type Unknown6 = string;
export type Unknown5 = Unknown6[];
/**
 * Description of the game
 */
export type Unknown7 = string;
/**
 * Manufacturer of the game
 */
export type Unknown8 = string;
/**
 * Year of the game
 */
export type Unknown9 = Unknown10 | Unknown11;
export type Unknown10 = number;
export type Unknown11 = string;
/**
 * URL of the homepage
 */
export type Unknown12 = string;
/**
 * URL of the github
 */
export type Unknown13 = string;
export type Unknown14 = string;
/**
 * Unique ID of the file. Must be unique per game.
 */
export type Unknown15 = number;
/**
 * File extension
 */
export type Unknown16 = string;
/**
 * File size in bytes
 */
export type Unknown17 = number;
/**
 * CRC32 checksum of the file
 */
export type Unknown18 = string;
/**
 * MD5 checksum of the file
 */
export type Unknown19 = string;
/**
 * SHA-1 checksum of the file
 */
export type Unknown20 = string;
/**
 * SHA-256 checksum of the file
 */
export type Unknown21 = string;
/**
 * ROM sources
 */
export type Files = File[];
export type Sources = Source[];
export type Games = Game[];

export interface GamesDb {
    games: Games;

    [k: string]: unknown;
}

export interface Game {
    name: Unknown;
    shortname?: Unknown1;
    region?: Unknown2;
    languages?: Unknown3;
    description?: Unknown7;
    manufacturer?: Unknown8;
    year?: Unknown9;
    links?: Links;
    sources: Sources;

    [k: string]: unknown;
}

export interface Links {
    homepage?: Unknown12;
    github?: Unknown13;

    [k: string]: Unknown14;
}

export interface Source {
    files?: Files;

    [k: string]: unknown;
}

export interface File {
    id: Unknown15;
    extension: Unknown16;
    size: Unknown17;
    crc32?: Unknown18;
    md5?: Unknown19;
    sha1?: Unknown20;
    sha256: Unknown21;

    [k: string]: unknown;
}